refac: layout.css.ts 생성및 레이아웃 전체 적용
