+++ components/ErrorBoundary/GlobalErrorBoundary.tsx
@@ -3,12 +3,17 @@
 import { ErrorBoundary } from '@suspensive/react';
 import { usePathname } from 'next/navigation';
 
+import { container } from '@/styles/layout.css';
+
 type Props = React.PropsWithChildren;
 
 function GlobalErrorBoundary({ children }: Props) {
   const pathname = usePathname();
   return (
-    <ErrorBoundary resetKeys={[pathname]} fallback={() => <div>ERROR</div>}>
+    <ErrorBoundary
+      resetKeys={[pathname]}
+      fallback={() => <div className={`${container}`}>ERROR</div>}
+    >
       {children}
     </ErrorBoundary>
   );
