+++ components/Hello/styles.css.ts
@@ -2,15 +2,6 @@ import { style } from '@vanilla-extract/css';
 
 import { vars } from '@/styles/theme.css';
 
-export const container = style({
-  width: '100%',
-  height: '100%',
-  display: 'flex',
-  flexDirection: 'column',
-  justifyContent: 'center',
-  alignItems: 'center',
-});
-
 export const button = style({
   padding: 4,
   background: 'none',
