+++ styles/layout.css.ts
@@ -1,0 +1,10 @@
+import { style } from '@vanilla-extract/css';
+
+export const container = style({
+  width: '100%',
+  height: '100%',
+  display: 'flex',
+  flexDirection: 'column',
+  justifyContent: 'center',
+  alignItems: 'center',
+});
